requirements.txt
streamlit>=1.28.0
deepl>=1.15.0
python-dotenv>=1.0.0
